% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.comparison_series_regression.R
\name{augment.comparison_series_regression}
\alias{augment.comparison_series_regression}
\title{Augment a comparison-series regression}
\usage{
\method{augment}{comparison_series_regression}(x, ...)
}
\arguments{
\item{x}{A `comparison_series_regression` object.}

\item{...}{Additional arguments. Currently unused.}
}
\value{
A tibble containing model data, fitted values and residuals.
}
\description{
Augment a comparison-series regression
}
