% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelate.R
\name{autocorrelate}
\alias{autocorrelate}
\title{Beregn autokorrelasjon i sammenligningsserier}
\usage{
autocorrelate(x, ...)
}
\arguments{
\item{x}{Objektet som skal analyseres.}

\item{...}{Tilleggsargumenter sendt til metode.}
}
\value{
Et autokorrelasjonsresultat. For `comparison_series` returneres
  et objekt av klassen `comparison_series_autocorrelation`.
}
\description{
Generisk funksjon for autokorrelasjonsanalyse. For et
`comparison_series` beregnes autokorrelasjon separat for hver serie
og de valgte lagene.
}
\seealso{
[combine_series()], [correlate()], [regress()]
}
