% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/business_cycle_explain.R
\name{business_cycle_explain}
\alias{business_cycle_explain}
\title{Forklar konjunkturklassifisering for ett år}
\usage{
business_cycle_explain(year, data = NULL, ...)
}
\arguments{
\item{year}{Året som skal forklares.}

\item{data}{Valgfritt NorMacro-datasett. Hvis `NULL`, brukes
standarddatasettet.}

\item{...}{Videre argumenter sendt til [business_cycle()].}
}
\value{
Resultatet for valgt år returneres usynlig som en tibble.
}
\description{
Viser konjunkturfase, samlet score, delkomponenter og underliggende
indikatorverdier for ett valgt år.
}
\examples{
\dontrun{
business_cycle_explain(
  2020
)
}

}
