% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_series.R
\name{compare_series}
\alias{compare_series}
\title{Sammenlign tidsserier}
\usage{
compare_series(
  variables,
  data = NULL,
  country = NULL,
  unit = NULL,
  base_year = NULL,
  normalize = TRUE,
  start_year = NULL,
  complete_cases = FALSE
)
}
\arguments{
\item{variables}{En tegnvektor med variabler som skal sammenlignes.}

\item{data}{Datasett som inneholder `Aar` og de valgte variablene.
Hvis `NULL`, brukes NorMacros standarddata.}

\item{country}{Valgfritt land ved analyse av internasjonale data.}

\item{unit}{Valgfri måleenhet.}

\item{base_year}{Valgfritt basisår ved normalisering.}

\item{normalize}{Logisk. Hvis `TRUE`, normaliseres seriene før
sammenligning.}

\item{start_year}{Valgfritt første år i sammenligningen.}

\item{complete_cases}{Logisk. Hvis `TRUE`, brukes bare år med komplette
observasjoner for alle valgte variabler.}
}
\value{
Et objekt som representerer de sammenlignede tidsseriene.
}
\description{
Sammenligner flere tidsserier i et felles datasett. Seriene kan
normaliseres til en felles skala, slik at utviklingen kan sammenlignes
selv når variablene har forskjellige måleenheter.
}
\examples{
compare_series(
  c("Inflasjon", "Styringsrente"),
  data = normacro_example
)

}
