% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Beregn korrelasjoner mellom sammenligningsserier}
\usage{
correlate(x, ...)
}
\arguments{
\item{x}{Objektet som skal analyseres.}

\item{...}{Tilleggsargumenter sendt til metode.}
}
\value{
Et korrelasjonsresultat. For `comparison_series` returneres et
  objekt av klassen `comparison_series_correlation`.
}
\description{
Generisk funksjon for korrelasjonsanalyse. For et `comparison_series`
beregnes parvise korrelasjoner mellom seriene.
}
\seealso{
[combine_series()], [regress()], [autocorrelate()]
}
