% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_explain.R
\name{episode_explain}
\alias{episode_explain}
\title{Forklar en historisk makroepisode}
\usage{
episode_explain(episode, data = NULL)
}
\arguments{
\item{episode}{Episode-ID fra [historical_episodes()].}

\item{data}{Norsk NorMacro-datasett. Hvis `NULL`, brukes [get_normacro()].}
}
\value{
Usynlig en liste med episodeinformasjon og datauttrekk.
}
\description{
Viser en kort beskrivelse av en kuratert historisk episode
og relevante NorMacro-variabler for perioden.
}
\examples{
episode_explain("finanskrisen_2008", data = normacro_example)

}
