% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_international_macro.R
\name{get_international_macro}
\alias{get_international_macro}
\title{Hent internasjonale makrodata}
\usage{
get_international_macro(export = FALSE, refresh = FALSE)
}
\arguments{
\item{export}{Logisk. Om datasettet også skal eksporteres til fil.}

\item{refresh}{Logisk. Om eksisterende cache skal fornyes.}
}
\value{
En tibble med internasjonale makroøkonomiske tidsserier.
}
\description{
Henter NorMacros standardiserte internasjonale makrodatasett.
}
\details{
Datasettet inneholder årlige makroøkonomiske indikatorer for Norge og
utvalgte europeiske land i et felles format med kolonnene `Aar` og
`Land`.
}
\examples{
\dontrun{
international <- get_international_macro()
}

}
