% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_international_metadata.R
\name{get_international_metadata}
\alias{get_international_metadata}
\title{Hent metadata for internasjonale NorMacro-data}
\usage{
get_international_metadata()
}
\value{
En tibble med internasjonale variabelmetadata.
}
\description{
Returnerer metadata for variablene i NorMacros internasjonale
makrodatasett.
}
\examples{
get_international_metadata()

}
