% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kostra_keyfigures.R
\name{get_kostra_keyfigures}
\alias{get_kostra_keyfigures}
\title{Hent utvalgte KOSTRA-nøkkeltall}
\usage{
get_kostra_keyfigures(regions, years = 2015:2025)
}
\arguments{
\item{regions}{En tegnvektor med KOSTRA-koder for kommuner eller andre
støttede enheter.}

\item{years}{År som skal hentes. Standard er 2015 til 2025.}
}
\value{
En tibble med standardiserte KOSTRA-nøkkeltall.
}
\description{
Henter og standardiserer utvalgte kommunale nøkkeltall fra
KOSTRA-tabell 12134.
}
\details{
Resultatet returneres i NorMacros standardiserte KOSTRA-format med
kolonnene `Enhet`, `Enhet_navn`, `Enhetstype` og `Aar`, i tillegg til
indikatorene i tabellen.
}
\examples{
\dontrun{
kostra <- get_kostra_keyfigures(
  regions = c("0301", "4601"),
  years = 2020:2025
)
}

}
