% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kostra_metadata.R
\name{get_kostra_metadata}
\alias{get_kostra_metadata}
\title{Hent metadata for et KOSTRA-datasett}
\usage{
get_kostra_metadata(data = NULL, table = NULL)
}
\arguments{
\item{data}{Valgfritt KOSTRA-datasett med attributtet `kostra_table`.}

\item{table}{Valgfritt KOSTRA-tabellnummer.}
}
\value{
En tibble med metadata for indikatorene i KOSTRA-tabellen.
}
\description{
Henter metadata for indikatorene i en støttet KOSTRA-tabell.
Tabellen kan identifiseres enten fra attributtene på et KOSTRA-datasett
eller ved å oppgi tabellnummer eksplisitt.
}
\examples{
get_kostra_metadata(
  data = normacro_kostra_example
)

}
