% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Hent metadata for et NorMacro-datasett}
\usage{
get_metadata(data = NULL)
}
\arguments{
\item{data}{Valgfritt NorMacro-datasett. Datastrukturen brukes til å
identifisere riktig metadataområde.}
}
\value{
En tibble med metadata for variablene i datasettet.
}
\description{
Returnerer metadata som passer til datasettet som oppgis. Dersom
`data` ikke oppgis, returneres metadata for de norske makrodataene.
}
\examples{
get_metadata()

get_metadata(
  normacro_international_example
)

}
