% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_normacro.R
\name{get_normacro}
\alias{get_normacro}
\title{Hent NorMacro-datasettet}
\usage{
get_normacro(export = FALSE)
}
\arguments{
\item{export}{Logisk verdi. Dersom `TRUE`, eksporteres NorMacro-datasettet
til `data_clean/normacro.csv` og `data_clean/normacro.rds`, og metadata
eksporteres til `data_clean/metadata_normacro.csv` og
`data_clean/metadata_normacro.xlsx`. Standard er `FALSE`.}
}
\value{
Et standardisert og validert NorMacro-datasett.
}
\description{
Bygger og returnerer NorMacros standardiserte makroøkonomiske datasett.
Datasettet valideres før det returneres.
}
\details{
Dersom `export = TRUE`, lagres datasettet og tilhørende metadata i
mappen `data_clean`.
}
\examples{
\dontrun{
data <- get_normacro()

# Hent data og lagre datasett og metadata lokalt
data <- get_normacro(
  export = TRUE
)
}

}
