% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth}
\alias{growth}
\title{Beregn vekst eller endring i sammenligningsserier}
\usage{
growth(x, ...)
}
\arguments{
\item{x}{Objektet som skal transformeres.}

\item{...}{Tilleggsargumenter sendt til metode.}
}
\value{
Et transformert objekt. For `comparison_series` returneres et
  nytt `comparison_series`-objekt.
}
\description{
Generisk funksjon for å beregne vekst eller absolutt endring i et
objekt. For `comparison_series` utføres transformasjonen separat for
hver serie.
}
\seealso{
[combine_series()], [index()], [normalize()]
}
