% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_table.R
\name{growth_table}
\alias{growth_table}
\title{Lag veksttabell}
\usage{
growth_table(variables, data = NULL, periods = c(1, 5, 10))
}
\arguments{
\item{variables}{En tegnvektor med variabler som skal analyseres.}

\item{data}{Datasett som inneholder `Aar` og de valgte variablene.
Hvis `NULL`, brukes NorMacros standarddata.}

\item{periods}{Numerisk vektor med antall år det skal beregnes vekst
over. Standard er 1, 5 og 10 år.}
}
\value{
En tibble med siste observasjon, samlet vekst og CAGR for hver
  angitt periode.
}
\description{
Beregner samlet vekst og gjennomsnittlig årlig vekst (CAGR) for
valgte variabler over én eller flere perioder.
}
\details{
Beregningene tar utgangspunkt i den siste tilgjengelige observasjonen
for hver variabel.
}
\examples{
growth_table(
  c("Befolkning", "Arbeidsstyrke", "Sysselsatte"),
  data = normacro_example
)

}
