% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Indekser en sammenligningsserie}
\usage{
index(x, ...)
}
\arguments{
\item{x}{Objektet som skal indekseres.}

\item{...}{Tilleggsargumenter sendt til metode.}
}
\value{
Et indeksert objekt.
}
\description{
Omdanner serier til en felles indeks med valgt basisår og basisverdi.
}
