% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_observations.R
\name{latest_observations}
\alias{latest_observations}
\title{Vis siste tilgjengelige observasjoner}
\usage{
latest_observations(data = NULL, category = NULL, source = NULL)
}
\arguments{
\item{data}{Datasettet som skal undersøkes. Hvis `NULL`, brukes
NorMacros standarddata.}

\item{category}{Valgfri kategori som resultatet skal begrenses til.}

\item{source}{Valgfri datakilde som resultatet skal begrenses til.}
}
\value{
En tibble med siste år, siste verdi og metadata for hver
  variabel.
}
\description{
Finner den siste tilgjengelige observasjonen for hver variabel i et
NorMacro-datasett og kombinerer resultatet med sentrale metadata.
}
\examples{
latest_observations(
  data = normacro_example
)

}
