% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_categories.R
\name{list_categories}
\alias{list_categories}
\title{List kategorier i NorMacro}
\usage{
list_categories(data = NULL, print = TRUE)
}
\arguments{
\item{data}{Valgfritt NorMacro-datasett. Hvis `NULL`, brukes standard
norske makrodata.}

\item{print}{Logisk. Hvis `TRUE`, skrives resultatet til konsollen.}
}
\value{
Resultatet returneres usynlig når det skrives til konsollen.
}
\description{
Viser hvilke tematiske kategorier som finnes i et NorMacro-datasett,
samt hvor mange variabler som tilhører hver kategori.
}
\examples{
list_categories(
  data = normacro_example
)

}
