% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{normacro_kostra_example}
\alias{normacro_kostra_example}
\title{KOSTRA-eksempeldatasett fra NorMacro}
\format{
En tibble med 18 observasjoner og 7 variabler:
\describe{
  \item{Enhet}{Kommunens KOSTRA-kode.}
  \item{Enhet_navn}{Kommunens navn.}
  \item{Enhetstype}{Type geografisk enhet.}
  \item{Aar}{År.}
  \item{Netto_driftsresultat}{Netto driftsresultat i prosent.}
  \item{Langsiktig_gjeld_uten_pensjonsforpliktelser}{
    Langsiktig gjeld uten pensjonsforpliktelser i prosent.
  }
  \item{Frie_inntekter_per_innbygger}{
    Frie inntekter per innbygger i kroner.
  }
}
}
\source{
NorMacro. Statisk eksempel basert på KOSTRA-tabell 12134.
}
\usage{
normacro_kostra_example
}
\description{
Et lite, statisk KOSTRA-datasett med utvalgte kommunale nøkkeltall
for Oslo, Bergen og Trondheim i perioden 2020–2025.
}
\details{
Datasettet er laget for bruk i eksempler, vignetter og dokumentasjon,
slik at KOSTRA-funksjonene kan demonstreres uten å hente data fra
eksterne API-er.

Datasettet representerer KOSTRA-tabell 12134, "Utvalgte nøkkeltall
for kommuneregnskap".


Datasettet har attributtene `dataset_type`, `kostra_table` og
`kostra_title`. Disse brukes av flere KOSTRA-funksjoner til å
identifisere datasettet og hente tilhørende metadata.
}
\keyword{datasets}
