% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normaliser sammenligningsserier}
\usage{
normalize(x, ...)
}
\arguments{
\item{x}{Objektet som skal normaliseres.}

\item{...}{Tilleggsargumenter sendt til metode.}
}
\value{
Et normalisert objekt. For `comparison_series` returneres et
  indeksert `comparison_series`-objekt med basisverdi 100.
}
\description{
Generisk funksjon for normalisering av serier. For `comparison_series`
tilsvarer normalisering indeksering til 100 i et felles basisår.
}
\seealso{
[combine_series()], [index()], [growth()]
}
