% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{overview_kostra_data}
\alias{overview_kostra_data}
\title{Gi oversikt over et KOSTRA-datasett}
\usage{
overview_kostra_data(data, print = TRUE)
}
\arguments{
\item{data}{Et standardisert KOSTRA-datasett.}

\item{print}{Logisk. Hvis `TRUE`, skrives en lesbar oversikt til
konsollen.}
}
\value{
Resultatet returneres usynlig. Når `print = TRUE`, skrives
  oversikten også til konsollen.
}
\description{
Oppsummerer strukturen i et standardisert KOSTRA-datasett og viser
blant annet tabellinformasjon, datadekning, antall enheter, antall
variabler og hvilke enhetstyper datasettet inneholder.
}
\examples{
overview_kostra_data(
  normacro_kostra_example
)

}
