% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kostra_benchmark.R
\name{plot_kostra_benchmark}
\alias{plot_kostra_benchmark}
\title{Visualiser KOSTRA-benchmark}
\usage{
plot_kostra_benchmark(variable, data, unit, year = NULL, descending = TRUE)
}
\arguments{
\item{variable}{Navnet på KOSTRA-indikatoren.}

\item{data}{Et KOSTRA-datasett.}

\item{unit}{KOSTRA-koden til enheten som skal fremheves.}

\item{year}{Valgfritt år. Hvis `NULL`, brukes siste tilgjengelige år.}

\item{descending}{Logisk. Hvis `TRUE`, rangeres høyeste verdi først.}
}
\value{
Et `ggplot`-objekt.
}
\description{
Visualiserer posisjonen til én KOSTRA-enhet i fordelingen av en valgt
indikator for ett år.
}
\details{
Figuren viser den valgte enheten sammen med median og intervallet
mellom første og tredje kvartil.
}
\examples{
plot_kostra_benchmark(
  "Netto_driftsresultat",
  data = normacro_kostra_example,
  unit = "0301",
  year = 2025
)

}
