% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kostra_ranking.R
\name{plot_kostra_ranking}
\alias{plot_kostra_ranking}
\title{Plot en KOSTRA-rangering}
\usage{
plot_kostra_ranking(
  variable,
  data,
  year = NULL,
  units = NULL,
  highlight = NULL,
  descending = TRUE
)
}
\arguments{
\item{variable}{Navnet på KOSTRA-indikatoren.}

\item{data}{Et KOSTRA-datasett.}

\item{year}{Valgfritt år. Hvis `NULL`, brukes siste tilgjengelige år.}

\item{units}{Valgfri tegnvektor med enhetskoder som skal inkluderes.}

\item{highlight}{Valgfri enhet som skal fremheves i figuren.}

\item{descending}{Logisk. Hvis `TRUE`, rangeres høyeste verdi først.}
}
\value{
Et `ggplot`-objekt.
}
\description{
Visualiserer rangeringen av KOSTRA-enheter etter en valgt indikator
i ett år.
}
\examples{
plot_kostra_ranking(
  "Netto_driftsresultat",
  data = normacro_kostra_example,
  year = 2025
)

}
