% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_series.R
\name{plot_series}
\alias{plot_series}
\title{Plot en tidsserie}
\usage{
plot_series(variable, data = NULL, metadata = NULL, countries = NULL)
}
\arguments{
\item{variable}{Navnet på variabelen som skal plottes.}

\item{data}{Datasett som inneholder `Aar` og variabelen som skal plottes.
Hvis `NULL`, hentes standarddatasettet med [get_normacro()].}

\item{metadata}{Metadata for datasettet. Hvis `NULL`, hentes metadata
automatisk.}

\item{countries}{Valgfri vektor med land som skal inkluderes når `data`
er et internasjonalt datasett.}
}
\value{
Et `ggplot`-objekt.
}
\description{
Lager en tidsseriefigur for en variabel i et NorMacro-datasett.
Funksjonen bruker metadata til å sette tittel, beskrivelse, måleenhet
og kilde når denne informasjonen er tilgjengelig.
}
\details{
Funksjonen støtter norske makrodata, internasjonale data med kolonnen
`Land` og KOSTRA-data med enhetsinformasjon.
}
\examples{
plot_series(
  "Inflasjon",
  data = normacro_example
)

}
