% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_kostra.R
\name{rank_kostra}
\alias{rank_kostra}
\title{Ranger KOSTRA-enheter}
\usage{
rank_kostra(variable, data, year = NULL, descending = TRUE, top_n = NULL)
}
\arguments{
\item{variable}{Navnet på KOSTRA-indikatoren.}

\item{data}{Et KOSTRA-datasett.}

\item{year}{Valgfritt år. Hvis `NULL`, brukes siste tilgjengelige år.}

\item{descending}{Logisk. Hvis `TRUE`, rangeres høyeste verdi først.}

\item{top_n}{Valgfritt antall øverste enheter som skal returneres.}
}
\value{
En tibble med rangering, enhetsinformasjon, år og verdi.
}
\description{
Rangerer kommuner eller andre KOSTRA-enheter etter verdien på en valgt
indikator i ett år.
}
\examples{
rank_kostra(
  "Netto_driftsresultat",
  data = normacro_kostra_example,
  year = 2025
)

}
