% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\title{Estimer regresjonsmodeller for sammenligningsserier}
\usage{
regress(x, ...)
}
\arguments{
\item{x}{Objektet som skal analyseres.}

\item{...}{Tilleggsargumenter sendt til metode.}
}
\value{
Et regresjonsresultat. For `comparison_series` returneres et
  objekt av klassen `comparison_series_regression`.
}
\description{
Generisk funksjon for regresjonsanalyse. For et `comparison_series`
angis modellen med en vanlig R-formel der variablene er `Serie_id`-ene
i objektet.
}
\seealso{
[combine_series()], [correlate()], [autocorrelate()]
}
