% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_normacro.R
\name{scale_colour_normacro}
\alias{scale_colour_normacro}
\title{NorMacro discrete colour scale}
\usage{
scale_colour_normacro(n = 7, ...)
}
\arguments{
\item{n}{Number of discrete series.}

\item{...}{Additional arguments passed to [ggplot2::scale_colour_manual()].}
}
\value{
A ggplot2 discrete colour scale.
}
\description{
Discrete colour scale used by NorMacro for categorical series.
}
\details{
The scale uses a colour-blind-friendly palette inspired by the
Okabe-Ito colour universal design palette. Up to eight distinct
series are supported.
}
