% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.normacro_report.R
\name{summary.normacro_report}
\alias{summary.normacro_report}
\title{Summarise a NorMacro macro report}
\usage{
\method{summary}{normacro_report}(object, ...)
}
\arguments{
\item{object}{A `normacro_report` object.}

\item{...}{Additional arguments.}
}
\value{
An object of class `summary_normacro_report`.
}
\description{
Summarise a NorMacro macro report
}
