% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_normacro.R
\name{theme_normacro}
\alias{theme_normacro}
\title{NorMacro plot theme}
\usage{
theme_normacro(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme object.
}
\description{
Standard ggplot2 theme used by NorMacro plots.
}
\details{
The theme is based on [ggplot2::theme_minimal()] with a small number
of adjustments for consistent titles, captions, grid lines and legends.
}
