% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelate.comparison_series.R
\name{autocorrelate.comparison_series}
\alias{autocorrelate.comparison_series}
\title{Beregn autokorrelasjon i sammenligningsserier}
\usage{
\method{autocorrelate}{comparison_series}(
  x,
  lags = 1:5,
  start_year = NULL,
  end_year = NULL,
  use = c("pairwise.complete.obs", "complete.obs", "everything", "all.obs",
    "na.or.complete"),
  method = c("pearson", "spearman", "kendall"),
  format = TRUE,
  ...
)
}
\arguments{
\item{x}{Et `comparison_series`-objekt.}

\item{lags}{Positive heltall som angir hvilke lag som skal beregnes.}

\item{start_year}{Valgfritt første år i analyseperioden.}

\item{end_year}{Valgfritt siste år i analyseperioden.}

\item{use}{Regel for håndtering av manglende observasjoner. Samme valg som
i `stats::cor()`.}

\item{method}{Korrelasjonsmetode: `"pearson"`, `"spearman"` eller
`"kendall"`.}

\item{format}{Logisk. Om resultatet skal formateres for utskrift.}

\item{...}{Videre argumenter til metoden.}
}
\value{
Et objekt av klassen `comparison_series_autocorrelation`.
}
\description{
Beregner autokorrelasjon separat for hver serie i et
`comparison_series`-objekt for ett eller flere lag.
}
