% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorMacro-package.R
\docType{package}
\name{NorMacro-package}
\alias{NorMacro}
\alias{NorMacro-package}
\title{NorMacro}
\description{
Makroøkonomiske tidsserier og analyseverktøy.
}
\author{
\strong{Maintainer}: Nils Kvilvang \email{nils@kvilvang.no}

Authors:
\itemize{
  \item Nils Kvilvang \email{nils@kvilvang.no}
}

}
\keyword{internal}
