% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_series.R
\name{scatter_series}
\alias{scatter_series}
\title{Lag scatterplott mellom to tidsserier}
\usage{
scatter_series(
  x,
  y,
  data = NULL,
  start_year = NULL,
  end_year = NULL,
  add_smooth = TRUE,
  label_years = FALSE,
  country = NULL,
  unit = NULL
)
}
\arguments{
\item{x}{Navnet på variabelen på x-aksen.}

\item{y}{Navnet på variabelen på y-aksen.}

\item{data}{Valgfritt NorMacro-datasett.}

\item{start_year}{Valgfritt første år i analyseperioden.}

\item{end_year}{Valgfritt siste år i analyseperioden.}

\item{add_smooth}{Logisk. Om en utjevnet trendlinje skal legges til.}

\item{label_years}{Logisk. Om observasjonene skal merkes med år.}

\item{country}{Valgfritt land når `data` inneholder internasjonale data.}

\item{unit}{Valgfri KOSTRA-enhet når `data` inneholder flere enheter.}
}
\value{
Et `ggplot`-objekt.
}
\description{
Visualiserer sammenhengen mellom to variabler i et NorMacro-datasett.
Figuren kan avgrenses til en bestemt periode og kan inkludere en
utjevnet trendlinje.
}
\details{
Funksjonen støtter norske makrodata, internasjonale data og KOSTRA-data.
}
\examples{
scatter_series(
  x = "BNP_Fastland_vekst",
  y = "Arbeidsledighetsrate_NAV",
  data = normacro_example
)

}
